==> Figure_3E-F_&_Figure_SI_4/0_CreateBins.ipynb <==
Creating new Master Peak file where peaks are divided into bins of 50 bp and the to total lenght of the peak is 5000 bp (2500 fome each side of the summit)

import pandas as pd 
import matplotlib as matplotlib
import pandas as pd
import numpy as np

import seaborn as sns
import matplotlib.pyplot as plt


master_peak=pd.read_csv("master_peakCleaned_60nt.bed", sep="\t", header=None)
master_peak.columns = ["chr", "start", "end"]
master_peak.head()



#master_peak.shape
len(master_peak)
master_peak

### adding a new column showing the peak number#
for i in range(1,1+len(master_peak)):
    master_peak.loc[i-1, "peak_number"] = f"peak_{i}"#i
master_peak



master_peak["peak_mid"] = np.ceil((master_peak["start"] + master_peak["end"])/2)
master_peak["peak_length"] = np.ceil((master_peak["end"] - master_peak["start"])/2)*2
master_peak.head()

### creating new dataframe with bin size of 50 and extetend length of peak by 2500bp on each side of the summit ###

#master_peak_bind=pd.DataFrame(columns=["chr", "start", "end", "peak_number", "peak_length"])

columns=["chr", "start", "end", "peak_number", "peak_length"]

master_peak_bin=pd.DataFrame(0, index=range(len(master_peak)*100),columns=columns)
master_peak_bin
#df = pd.DataFrame(0, index=range(24897), columns=columns)


rows = []
for i in range(1, 1 + len(master_peak)):
    for j in range(-2500, 2500, 50):
        rows.append({"chr": master_peak.loc[i-1, "chr"],
                     "start": master_peak.loc[i-1, "peak_mid"] + j - 25,
                     "end": master_peak.loc[i-1, "peak_mid"] + j + 25,
                     "peak_number": master_peak.loc[i-1, "peak_number"],
                     "peak_length": 50})

# Convert list to DataFrame once
master_peak_bind = pd.DataFrame(rows)


master_peak_bind

# make master_peak_bind[["start", "end"]] as type int
master_peak_bind[["start", "end"]] = master_peak_bind[["start", "end"]].astype(int)

# remove chrUn_GL456370 and chrM
master_peak_bind = master_peak_bind[master_peak_bind["chr"] != "chrUn_GL456370"]
master_peak_bind = master_peak_bind[master_peak_bind["chr"] != "chrM"]

# save master_peak_bind to a tab separated file without the index and  header
master_peak_bind[["chr", "start", "end"]].to_csv("master_peakCleaned_60nt_binned.bed", sep="\t", index=False, header=False)


# test if start or end is negative
master_peak_bind[master_peak_bind["start"] < 0]['chr'].value_counts()
