==> Figure_3E-F_&_Figure_SI_4/2_count_binned_zscore.ipynb <==
# Main analysis

from count matrix 
zscore

# create count Matrix

this code generates for all conditions+TF a seperate count matrix for the peask survived after z-score filtering !


#imports

import pandas as pd
import numpy as np
import glob
import os


# creating count Matrix

# seperate condition and region
conditions = ["MYC_ESC", "MYC_EpiLC", "NANOG_ESC", "NANOG_EpiLC", "YAP1_ESC", "YAP1_EpiLC", "OCT4_ESC", "OCT4_EpiLC"]

# Path to the folder containing the .txt files
input_folder = "./txt"

# List all .txt files that start with "R"
file_list = glob.glob(os.path.join(input_folder, "R*.txt"))

# Initialize an empty list to store individual DataFrames
#dfs = []

# interate through each condition
index_file = 0
for condition in conditions:

    # Initialize an empty list to store individual DataFrames
    dfs = []
    
    # Iterate through each file
    for file in file_list:
        #print(file)
        # Extract the file name without the extension (used as the column name)
        column_name = os.path.splitext(os.path.basename(file))[0]

        # if the file name contains the condition
        if condition in column_name:
            index_file += 1
            print(index_file, column_name)
            # Read the file into a DataFrame
            df = pd.read_csv(file, sep="\t", header=None, names=[column_name], index_col=0)
            #display(df)
            
            # Append the DataFrame to the list
            dfs.append(df)
        
       
    # concatenate all the dataframes in dfs list
    count_matrix = pd.concat(dfs, axis=1)
    #display(count_matrix)

    output_file = "./merged_count_" + condition + "_binned.txt"
    count_matrix.to_csv(output_file, sep="\t")
