==> Figure_3E-F_&_Figure_SI_4/4_peaks_significant_zscore.ipynb <==
import numpy as np
import matplotlib.pyplot as plt
import pandas as pd
import seaborn as sns


MYC_zscores = pd.read_csv("master_peak_ZSCOREresults_MYC_ESC_EpiLC_60nt_REVIEW_NORMALIZED.txt", sep='\t') #, index_col=0
NANOG_zscores = pd.read_csv("master_peak_ZSCOREresults_NANOG_ESC_EpiLC_60nt_REVIEW_NORMALIZED.txt", sep='\t') #, index_col=0
OCT4_zscores = pd.read_csv("master_peak_ZSCOREresults_OCT4_ESC_EpiLC_60nt_REVIEW_NORMALIZED.txt", sep='\t') #, index_col=0

MYC_zscores['region'] = MYC_zscores['chr'] + ':' + MYC_zscores['start'].astype(str) + '-' + MYC_zscores['end'].astype(str)
NANOG_zscores['region'] = NANOG_zscores['chr'] + ':' + NANOG_zscores['start'].astype(str) + '-' + NANOG_zscores['end'].astype(str)
OCT4_zscores['region'] = OCT4_zscores['chr'] + ':' + OCT4_zscores['start'].astype(str) + '-' + OCT4_zscores['end'].astype(str)

#MYC_zscores.head()

MYC_zscores_pvals=MYC_zscores[['region','z-score','adjusted p-value']]
NANOG_zscores_pvals=NANOG_zscores[['region','z-score','adjusted p-value']]
OCT4_zscores_pvals=OCT4_zscores[['region','z-score','adjusted p-value']]
#MYC_zscores_pvals.head()
NANOG_zscores_pvals.head()

MYC_zscores_pvals.set_index('region', inplace=True)
NANOG_zscores_pvals.set_index('region', inplace=True)
OCT4_zscores_pvals.set_index('region', inplace=True)
#MYC_zscores_pvals.head()

cutoff = 0.001 #0.01 

print(MYC_zscores_pvals.shape)
print(NANOG_zscores_pvals.shape)
print(OCT4_zscores_pvals.shape)

MYC_zscores_pvals_sig = MYC_zscores_pvals[MYC_zscores_pvals['adjusted p-value'] <= cutoff]
NANOG_zscores_pvals_sig = NANOG_zscores_pvals[NANOG_zscores_pvals['adjusted p-value'] <= cutoff]
OCT4_zscores_pvals_sig = OCT4_zscores_pvals[OCT4_zscores_pvals['adjusted p-value'] <= cutoff]

print(MYC_zscores_pvals_sig.shape)
print(NANOG_zscores_pvals_sig.shape)
print(OCT4_zscores_pvals_sig.shape)

display(MYC_zscores_pvals_sig.head())

#NANOG_zscores_pvals_sig

# 0.01
#(10048, 2)
#(10328, 2)
#(11041, 2)

# 0.001
(9660, 2)
(9842, 2)
(10811, 2)




MYC_zscores_pvals_sig.to_csv('MYC_zscores_pvals_sig_REVIEW_NORMALIZED.txt', sep='\t')
NANOG_zscores_pvals_sig.to_csv('NANOG_zscores_pvals_sig_REVIEW_NORMALIZED.txt', sep='\t')
OCT4_zscores_pvals_sig.to_csv('OCT4_zscores_pvals_sig_REVIEW_NORMALIZED.txt', sep='\t')
